Refactor large modules into packages
