+++ TrinityBackendFastAPI/app/DataStorageRetrieval/db/connection.py
@@ -1,0 +1,11 @@
+import os
+
+try:
+    import asyncpg  # type: ignore
+except ModuleNotFoundError:  # pragma: no cover - optional dependency
+    asyncpg = None
+
+POSTGRES_HOST = os.getenv("POSTGRES_HOST", "postgres")
+POSTGRES_DB = os.getenv("POSTGRES_DB", "trinity_db")
+POSTGRES_USER = os.getenv("POSTGRES_USER", "trinity_user")
+POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD", "trinity_pass")
